move UserBadge, update Nameplate description
