+++ webui/react/src/hooks/useModal/Workspace/useModalWorkspaceAddMember.tsx
@@ -4,7 +4,7 @@ import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
 import Form from 'components/kit/Form';
 import Nameplate from 'components/kit/Nameplate';
-import UserBadge from 'components/kit/UserBadge';
+import UserBadge from 'components/UserBadge';
 import { assignRolesToGroup, assignRolesToUser } from 'services/api';
 import { V1Group, V1Role } from 'services/api-ts-sdk';
 import Icon from 'shared/components/Icon';
